+++ designer/client/src/components/Process/ProcessBackButton.tsx
@@ -4,14 +4,14 @@ import ProcessBackIcon from "../../assets/img/arrows/back-process.svg";
 import { useTranslation } from "react-i18next";
 import { matchPath, useLocation } from "react-router-dom";
 import { MetricsBasePath } from "../../containers/paths";
-import styled from "@emotion/styled";
+import { styled } from "@mui/material";
 import { variables } from "../../stylesheets/variables";
 
 const BackIcon = styled(ProcessBackIcon)(() => ({
     height: "12px",
 }));
 
-const ButtonText = styled.span(() => ({
+const ButtonText = styled("span")(() => ({
     fontSize: "14px",
     fontWeight: 600,
     color: variables.buttonTextColor,
@@ -27,7 +27,6 @@ const ProcessLinkButton = styled(ProcessLink)(() => ({
     alignItems: "center",
     padding: "0 8px",
     cursor: "pointer",
-
     "&:hover, &:focus": {
         backgroundColor: variables.menuButtonActiveBKColor,
     },