+++ designer/client/src/components/graph/Graph.tsx
@@ -4,32 +4,32 @@ import { cloneDeep, debounce, isEmpty, isEqual, keys, sortBy, without } from "lo
 import React from "react";
 import { findDOMNode } from "react-dom";
 import "../../stylesheets/graph.styl";
-import { filterDragHovered, getLinkNodes, setLinksHovered } from "./dragHelpers";
+import { filterDragHovered, getLinkNodes, setLinksHovered } from "./utils/dragHelpers";
 import { updateNodeCounts } from "./EspNode/element";
 import { GraphPaperContainer } from "./focusable";
 import { applyCellChanges, calcLayout, createPaper, isModelElement } from "./GraphPartialsInTS";
 import styles from "./graphTheme.styl";
-import { Events } from "./joint-events";
+import { Events, GraphProps } from "./types";
 import NodeUtils from "./NodeUtils";
 import { PanZoomPlugin } from "./PanZoomPlugin";
 import { RangeSelectedEventData, RangeSelectPlugin, SelectionMode } from "./RangeSelectPlugin";
 import "./svg-export/export.styl";
 import { prepareSvg } from "./svg-export/prepareSvg";
-import * as GraphUtils from "./GraphUtils";
+import * as GraphUtils from "./utils/graphUtils";
 import { ComponentDragPreview } from "../ComponentDragPreview";
 import { rafThrottle } from "./rafThrottle";
 import { isEdgeEditable } from "../../common/EdgeUtils";
 import { NodeId, NodeType, Process, ProcessDefinitionData } from "../../types";
 import { Layout, NodePosition, Position } from "../../actions/nk";
 import { UserSettings } from "../../reducers/userSettings";
-import { GraphProps } from "./GraphWrapped";
 import User from "../../common/models/User";
 import { updateLayout } from "./GraphPartialsInTS/updateLayout";
 import { getDefaultLinkCreator } from "./EspNode/link";
 import ProcessUtils from "../../common/ProcessUtils";
-import { isTouchDevice, isTouchEvent } from "../../helpers/detectDevice";
+import { isTouchDevice, isTouchEvent, LONG_PRESS_TIME } from "../../helpers/detectDevice";
 import { batchGroupBy } from "../../reducers/graph/batchGroupBy";
 import { createUniqueArrowMarker } from "./arrowMarker";
+import { cx } from "@emotion/css";
 
 interface Props extends GraphProps {
     processCategory: string;
@@ -41,9 +41,11 @@ interface Props extends GraphProps {
     isPristine?: boolean;
 }
 
-function handleActionOnLongPress(
-    shortPressAction: (cellView: dia.CellView, event: dia.Event) => void,
-    longPressAction: (cellView: dia.CellView, event: dia.Event) => void,
+function handleActionOnLongPress<T extends dia.CellView>(
+    shortPressAction: ((cellView: T, event: dia.Event) => void) | null,
+    longPressAction: (cellView: T, event: dia.Event) => void,
+    getPinchEventActive: () => boolean,
+    longPressTime = LONG_PRESS_TIME,
 ) {
     let pressTimer;
 
@@ -51,25 +53,31 @@ function handleActionOnLongPress(
         clearTimeout(pressTimer);
     };
 
-    return (cellView: dia.CellView, evt: dia.Event) => {
-        const { paper, model } = cellView;
+    return (cellView: T, evt: dia.Event) => {
+        const { paper } = cellView;
 
         // let's clear all pointer click events on start
-        paper.off(Events.CELL_POINTERCLICK, shortPressAction);
-        paper.on(Events.CELL_POINTERCLICK, shortPressAction);
-        paper.on(Events.CELL_POINTERUP, releasePress);
+        if (shortPressAction) {
+            paper.off(Events.CELL_POINTERCLICK, shortPressAction);
+            paper.once(Events.CELL_POINTERCLICK, shortPressAction);
+        }
 
-        const LONG_PRESS_TIME = 500;
-
-        // Discard specific events on long press action
-        model.once(Events.CHANGE_POSITION, releasePress);
-        model.once(Events.CELL_POINTERUP, releasePress);
+        // Discard long press action on specific events
+        paper.once(Events.CELL_POINTERUP, releasePress);
+        paper.once(Events.CELL_POINTERMOVE, releasePress);
 
         pressTimer = window.setTimeout(() => {
             // Stop single click event when longPress fired
-            paper.off(Events.CELL_POINTERCLICK, shortPressAction);
+            if (shortPressAction) {
+                paper.off(Events.CELL_POINTERCLICK, shortPressAction);
+            }
+
+            if (getPinchEventActive()) {
+                return;
+            }
+
             longPressAction(cellView, evt);
-        }, LONG_PRESS_TIME);
+        }, longPressTime);
     };
 }
 
@@ -245,7 +253,7 @@ export class Graph extends React.Component<Props> {
             }
         };
 
-        const selectNode = (cellView, evt) => {
+        const selectNode = (cellView: dia.CellView, evt: dia.Event) => {
             if (this.props.nodeSelectionEnabled) {
                 const nodeDataId = cellView.model.attributes.nodeData?.id;
                 if (!nodeDataId) {
@@ -270,7 +278,14 @@ export class Graph extends React.Component<Props> {
         };
 
         if (isTouchDevice()) {
-            this.processGraphPaper.on(Events.CELL_POINTERDOWN, handleActionOnLongPress(showNodeDetails, selectNode));
+            this.processGraphPaper.on(
+                Events.CELL_POINTERDOWN,
+                handleActionOnLongPress(showNodeDetails, selectNode, this.panAndZoom.getPinchEventActive),
+            );
+            this.processGraphPaper.on(
+                Events.LINK_POINTERDOWN,
+                handleActionOnLongPress(null, ({ model }) => model.remove(), this.panAndZoom.getPinchEventActive, LONG_PRESS_TIME * 1.5),
+            );
         } else {
             this.processGraphPaper.on(Events.CELL_POINTERCLICK, selectNode);
             this.processGraphPaper.on(Events.CELL_POINTERDBLCLICK, showNodeDetails);
@@ -290,7 +305,7 @@ export class Graph extends React.Component<Props> {
         this.panAndZoom = new PanZoomPlugin(this.processGraphPaper);
 
         if (this.props.nodeSelectionEnabled) {
-            new RangeSelectPlugin(this.processGraphPaper);
+            new RangeSelectPlugin(this.processGraphPaper, this.panAndZoom.getPinchEventActive);
             this.processGraphPaper.on("rangeSelect:selected", ({ elements, mode }: RangeSelectedEventData) => {
                 const nodes = elements.filter((el) => isModelElement(el)).map(({ id }) => id.toString());
                 if (mode === SelectionMode.toggle) {
@@ -584,6 +599,7 @@ export class Graph extends React.Component<Props> {
         return (
             <>
                 <GraphPaperContainer
+                    className={cx({ touch: isTouchDevice() })}
                     ref={this.setEspGraphRef}
                     onResize={isFragment ? () => this.panAndZoom.fitSmallAndLargeGraphs() : null}
                     id={divId}