+++ designer/client/src/containers/Page.tsx
@@ -1,5 +1,3 @@
-import React from "react";
-import "../stylesheets/visualization.styl";
 import styled from "@emotion/styled";
 
 export const Page = styled.div({