+++ security/src/main/scala/pl/touk/nussknacker/ui/security/basicauth/BasicAuthenticationResources.scala
@@ -1,19 +1,37 @@
 package pl.touk.nussknacker.ui.security.basicauth
 
 import akka.http.scaladsl.server.directives.{AuthenticationDirective, SecurityDirectives}
-import pl.touk.nussknacker.ui.security.api.{AnonymousAccess, AuthenticatedUser, AuthenticationResources, FrontendStrategySettings}
+import pl.touk.nussknacker.ui.security.api._
+import sttp.model.headers.WWWAuthenticateChallenge
+import sttp.tapir.EndpointInput.Auth
+import sttp.tapir._
 
-class BasicAuthenticationResources(realm: String, configuration: BasicAuthenticationConfiguration) extends AuthenticationResources with AnonymousAccess {
+import scala.concurrent.{ExecutionContext, Future}
+
+class BasicAuthenticationResources(realm: String, configuration: BasicAuthenticationConfiguration)
+                                  (implicit executionContext: ExecutionContext)
+  extends AuthenticationResources
+    with AnonymousAccess {
+
   val name: String = configuration.name
 
   val frontendStrategySettings: FrontendStrategySettings = FrontendStrategySettings.Browser
 
   val anonymousUserRole: Option[String] = configuration.anonymousUserRole
 
+  private val authenticator = BasicHttpAuthenticator(configuration)
+
   def authenticateReally(): AuthenticationDirective[AuthenticatedUser] =
     SecurityDirectives.authenticateBasicAsync(
-      authenticator = BasicHttpAuthenticator(configuration),
+      authenticator = authenticator,
       realm = realm
     )
 
+  override def authenticationMethod(): Auth[AuthCredentials, _] = {
+    auth.basic[AuthCredentials](WWWAuthenticateChallenge.basic.realm(realm))
+  }
+
+  override def authenticate(authCredentials: AuthCredentials): Future[Option[AuthenticatedUser]] = {
+    authenticator.authenticate(authCredentials)
+  }
 }