+++ security/src/main/scala/pl/touk/nussknacker/ui/security/oidc/OidcAuthenticationConfiguration.scala
@@ -1,7 +1,10 @@
 package pl.touk.nussknacker.ui.security.oidc
 
 import com.typesafe.config.Config
+import net.ceedubs.ficus.readers.ValueReader
+import pl.touk.nussknacker.engine.util.config.FicusReaders.forDecoder
 import pl.touk.nussknacker.engine.util.config.URIExtensions
+import pl.touk.nussknacker.ui.security.api.{AuthenticationConfiguration, FrontendStrategySettings}
 import pl.touk.nussknacker.ui.security.oauth2.ProfileFormat.OIDC
 import pl.touk.nussknacker.ui.security.oauth2.UsernameClaim.UsernameClaim
 import pl.touk.nussknacker.ui.security.oauth2.{JwtConfiguration, OAuth2Configuration, TokenCookieConfig}
@@ -33,6 +36,7 @@ case class OidcAuthenticationConfiguration(usersFile: URI,
                                            tokenCookie: Option[TokenCookieConfig] = None,
                                            usernameClaim: Option[UsernameClaim] = None,
                                            accessTokenIsJwt: Boolean = false,
+                                           overrideFrontendAuthenticationStrategy: Option[FrontendStrategySettings] = None,
                                           ) extends URIExtensions {
 
   lazy val oAuth2Configuration: OAuth2Configuration = OAuth2Configuration(
@@ -65,6 +69,7 @@ case class OidcAuthenticationConfiguration(usersFile: URI,
     accessTokenRequestContentType = MediaType.ApplicationXWwwFormUrlencoded.toString(),
     anonymousUserRole = anonymousUserRole,
     tokenCookie = tokenCookie,
+    overrideFrontendAuthenticationStrategy = overrideFrontendAuthenticationStrategy,
     usernameClaim = usernameClaim
   )
 
@@ -86,13 +91,13 @@ case class OidcAuthenticationConfiguration(usersFile: URI,
 
 object OidcAuthenticationConfiguration {
 
-  import net.ceedubs.ficus.readers.ArbitraryTypeReader._
-  import pl.touk.nussknacker.engine.util.config.CustomFicusInstances._
-  import pl.touk.nussknacker.ui.security.api.AuthenticationConfiguration.authenticationConfigPath
-  import net.ceedubs.ficus.readers.EnumerationReader._
-
-  def create(config: Config): OidcAuthenticationConfiguration =
-    config.as[OidcAuthenticationConfiguration](authenticationConfigPath)
+  def create(config: Config): OidcAuthenticationConfiguration = {
+    import net.ceedubs.ficus.readers.ArbitraryTypeReader._
+    import net.ceedubs.ficus.readers.EnumerationReader._
+    import pl.touk.nussknacker.engine.util.config.CustomFicusInstances._
+    implicit val valueReader: ValueReader[FrontendStrategySettings] = forDecoder
+    config.as[OidcAuthenticationConfiguration](AuthenticationConfiguration.authenticationConfigPath)
+  }
 
   def createWithDiscovery(config: Config)(implicit ec: ExecutionContext, sttpBackend: SttpBackend[Future, Any]): OidcAuthenticationConfiguration =
     create(config).withDiscovery