+++ security/src/test/scala/pl/touk/nussknacker/ui/security/oauth2/OAuth2AuthenticationResourcesSpec.scala
@@ -4,9 +4,11 @@ import akka.http.scaladsl.model.StatusCodes
 import akka.http.scaladsl.model.headers.{HttpCookie, `Set-Cookie`}
 import akka.http.scaladsl.testkit.ScalatestRouteTest
 import de.heikoseeberger.akkahttpcirce.FailFastCirceSupport
-
+import io.circe.Json
+import org.scalatest.EitherValues
 import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.should.Matchers
+import pl.touk.nussknacker.ui.security.api.FrontendStrategySettings
 import pl.touk.nussknacker.ui.security.http.RecordingSttpBackend
 import sttp.client3.Response
 import sttp.client3.testing.SttpBackendStub
@@ -15,7 +17,12 @@ import sttp.model.{StatusCode, Uri}
 import scala.concurrent.Future
 import scala.language.higherKinds
 
-class OAuth2AuthenticationResourcesSpec extends AnyFunSpec with Matchers with ScalatestRouteTest with FailFastCirceSupport {
+class OAuth2AuthenticationResourcesSpec
+  extends AnyFunSpec
+    with Matchers
+    with ScalatestRouteTest
+    with FailFastCirceSupport
+    with EitherValues {
 
   private val defaultConfig = ExampleOAuth2ServiceFactory.testConfig
 
@@ -80,9 +87,12 @@ class OAuth2AuthenticationResourcesSpec extends AnyFunSpec with Matchers with Sc
     authenticationOauth2(authenticationResources(), authorizationCode) ~> check {
       status shouldBe StatusCodes.OK
       header[`Set-Cookie`] shouldBe None
-      val response = responseAs[Oauth2AuthenticationResponse]
-      response.accessToken shouldEqual accessToken
-      response.tokenType shouldEqual "Bearer"
+      val response = responseAs[Json]
+
+      response.hcursor.downField("accessToken").as[String].value should be (accessToken)
+      response.hcursor.downField("access_token").as[String].value should be (accessToken)
+      response.hcursor.downField("tokenType").as[String].value should be ("Bearer")
+      response.hcursor.downField("token_type").as[String].value should be ("Bearer")
     }
   }
 
@@ -93,4 +103,26 @@ class OAuth2AuthenticationResourcesSpec extends AnyFunSpec with Matchers with Sc
       header[`Set-Cookie`] shouldBe Some(`Set-Cookie`(HttpCookie(name = cookieConfig.name, value = accessToken, httpOnly = true, secure = true, path = cookieConfig.path)))
     }
   }
+
+  it("should return authentication settings") {
+    Get(s"/authentication/oauth2/settings") ~> routes(authenticationResources()) ~> check {
+      status shouldBe StatusCodes.OK
+      val response = responseAs[Json]
+      response.hcursor.downField("authorizeUrl").as[String].value shouldBe s"${defaultConfig.authorizeUri}?client_id=${defaultConfig.clientId}"
+      response.hcursor.downField("jwtIdTokenNonceVerificationRequired").as[Boolean].value shouldBe false
+      response.hcursor.downField("implicitGrantEnabled").as[Boolean].value shouldBe false
+      response.hcursor.downField("anonymousAccessAllowed").as[Boolean].value shouldBe false
+      response.hcursor.downField("strategy").as[String].value shouldBe "OAuth2"
+    }
+  }
+
+  it("should return overriden authentication settings") {
+    val frontendSettings = FrontendStrategySettings.Remote("http://some.remote.url")
+    Get(s"/authentication/oauth2/settings") ~> routes(authenticationResources(defaultConfig.copy(overrideFrontendAuthenticationStrategy = Some(frontendSettings)))) ~> check {
+      status shouldBe StatusCodes.OK
+      val response = responseAs[Json]
+      response.hcursor.downField("moduleUrl").as[String].value shouldBe frontendSettings.moduleUrl
+    }
+  }
+
 }